Separate class and interface tests
